+++ src/gwt/src/org/rstudio/studio/client/RStudioGinjector.java
@@ -31,6 +31,7 @@ import org.rstudio.core.client.theme.WindowFrame;
 import org.rstudio.core.client.widget.CaptionWithHelp;
 import org.rstudio.core.client.widget.LocalRepositoriesWidget;
 import org.rstudio.core.client.widget.ModifyKeyboardShortcutsWidget;
+import org.rstudio.core.client.widget.RStudioThemedFrame;
 import org.rstudio.studio.client.application.Application;
 import org.rstudio.studio.client.application.ApplicationInterrupt;
 import org.rstudio.studio.client.application.events.EventBus;
@@ -107,6 +108,7 @@ import org.rstudio.studio.client.workbench.views.source.editors.explorer.view.Ob
 import org.rstudio.studio.client.workbench.views.source.editors.explorer.view.ObjectExplorerEditingTargetWidget;
 import org.rstudio.studio.client.workbench.views.source.editors.text.AceEditor;
 import org.rstudio.studio.client.workbench.views.source.editors.text.AceEditorIdleCommands;
+import org.rstudio.studio.client.workbench.views.source.editors.text.AceEditorMixins;
 import org.rstudio.studio.client.workbench.views.source.editors.text.TextEditingTargetIdleMonitor;
 import org.rstudio.studio.client.workbench.views.source.editors.text.AceEditorWidget;
 import org.rstudio.studio.client.workbench.views.source.editors.text.ChunkSatellite;
@@ -231,6 +233,8 @@ public interface RStudioGinjector extends Ginjector
    void injectMembers(ObjectExplorerEditingTargetWidget widget);
    void injectMembers(ObjectExplorerDataGrid widget);
    void injectMembers(TerminalInfoDialog infoDialog);
+   void injectMembers(AceEditorMixins mixins);
+   void injectMembers(RStudioThemedFrame frame);
    
    public static final RStudioGinjector INSTANCE = GWT.create(RStudioGinjector.class);